worked on scoring
git-svn-id: https://guitarstorm.googlecode.com/svn/trunk@28 9eb3be25-a147-0410-a550-09e9c04672db

+++ Capture.h
@@ -48,7 +48,7 @@ public:
 
 	void stop();	// stop capture
 
-	float* readLast(int l);	// get the last l samples captured
+	double* readLast(int l);	// get the last l samples captured
 
 protected:
     PaStream *stream;
